ai-service/requirements.txt
# File:    ai-service/requirements.txt
# Purpose: Python dependencies for the FastAPI AI service.
# Owner:   Navanish

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM and Embeddings (Gemini)
google-genai>=1.0.0

# Database and Vector Search
psycopg[binary]>=3.1.18
pgvector>=0.3.2
sqlalchemy>=2.0.23

# Document Processing
pypdf>=4.0.0
python-multipart>=0.0.9

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.7.0
httpx>=0.27.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.0

# Auth
python-jose[cryptography]>=3.3.0
